+++ pcap4j-sample/src/main/java/org/pcap4j/sample/IcmpV4ErrReplyer.java
@@ -170,7 +170,7 @@ public class IcmpV4ErrReplyer {
               return;
             }
 
-            EthernetPacket.Builder eb1 = (EthernetPacket.Builder) packet.getBuilder();
+            EthernetPacket.Builder eb1 = packet.getBuilder().get(EthernetPacket.Builder.class);
             ArpPacket.Builder ab = eb1.get(ArpPacket.Builder.class);
 
             ab.srcHardwareAddr(MAC_ADDR)