+++ CHANGES.md
@@ -1,6 +1,7 @@
 Future
 ======
 ### New Features ###
+* [PR#146](https://github.com/kaitoy/pcap4j/pull/146): Add PcapPacket builder. `PcapPacket#getBuilder()` is changed such that it returns a PcapPacket builder instance instead of payload's builder. 
 
 ### Bug Fixes ###